Linker errors fixed after very long debugging... Restructured project
