+++ RRTracer/RRTracer.vcxproj.filters
@@ -13,38 +13,41 @@
       <UniqueIdentifier>{67DA6AB6-F800-4c08-8B7A-83BB121AAD01}</UniqueIdentifier>
       <Extensions>rc;ico;cur;bmp;dlg;rc2;rct;bin;rgs;gif;jpg;jpeg;jpe;resx;tiff;tif;png;wav;mfcribbon-ms</Extensions>
     </Filter>
-    <Filter Include="Tuple Files">
-      <UniqueIdentifier>{6d633b0b-2195-4d46-ba6c-9bfbe11d6e5e}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="Color Files">
-      <UniqueIdentifier>{48fdb880-aa88-4a1e-92dc-adc986e1edcf}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="Canvas Files">
-      <UniqueIdentifier>{03cb2e2f-957d-4ccb-8d10-60000590d946}</UniqueIdentifier>
-    </Filter>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="Utilities.hpp">
-      <Filter>Source Files</Filter>
+    <ClInclude Include="Color.h">
+      <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="Tuple.hpp">
-      <Filter>Tuple Files</Filter>
+    <ClInclude Include="Tuple.h">
+      <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="TupleFactory.hpp">
-      <Filter>Tuple Files</Filter>
+    <ClInclude Include="TupleFactory.h">
+      <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="TupleUtilities.hpp">
-      <Filter>Tuple Files</Filter>
+    <ClInclude Include="TupleUtilities.h">
+      <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="Canvas.hpp">
-      <Filter>Canvas Files</Filter>
+    <ClInclude Include="Utilities.h">
+      <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="Color.hpp">
-      <Filter>Color Files</Filter>
+    <ClInclude Include="Canvas.h">
+      <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="main.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="Color.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="Tuple.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="TupleFactory.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="Canvas.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
