+++ RRTracer/TupleUtilities.h
@@ -1,11 +1,11 @@
-#ifndef TUPLE_UTILITIES
-#define TUPLE_UTILITIES
+#ifndef RRTTUPLEUTILS_TUPLEUTILITIES_HPP
+#define RRTTUPLEUTILS_TUPLEUTILITIES_HPP
 
-#include "Tuple.hpp"
+#include "Tuple.h"
 
 namespace RRTTupleUtils
 {
-	static RRT::Tuple Normalize(const RRT::Tuple& tuple)
+	inline RRT::Tuple Normalize(const RRT::Tuple& tuple)
 	{
 		float magnitude = tuple.magnitude();
 		float x_norm = tuple.X() / magnitude;
@@ -13,14 +13,14 @@ namespace RRTTupleUtils
 		float z_norm = tuple.Z() / magnitude;
 		return { x_norm, y_norm, z_norm, tuple.W() };
 	}
-	
-	static float Dot(const RRT::Tuple& tuple_one, const RRT::Tuple& tuple_two)
+
+	inline float Dot(const RRT::Tuple& tuple_one, const RRT::Tuple& tuple_two)
 	{
 		float dot_product = tuple_one.X() * tuple_two.X() + tuple_one.Y() * tuple_two.Y() + tuple_one.Z() * tuple_two.Z() + tuple_one.W() * tuple_two.W();
 		return dot_product;
 	}
 
-	static RRT::Tuple Cross(const RRT::Tuple& tuple_one, const RRT::Tuple& tuple_two)
+	inline RRT::Tuple Cross(const RRT::Tuple& tuple_one, const RRT::Tuple& tuple_two)
 	{
 		float y_z = tuple_one.Y() * tuple_two.Z() - tuple_one.Z() * tuple_two.Y();
 		float z_x = tuple_one.Z() * tuple_two.X() - tuple_one.X() * tuple_two.Z();
@@ -30,4 +30,4 @@ namespace RRTTupleUtils
 	}
 }
 
-#endif // !TUPLE_UTILITIES
+#endif // !RRTTUPLEUTILS_TUPLEUTILITIES_HPP
