+++ RRTracer/main.cpp
@@ -2,7 +2,6 @@
 
 int main()
 {
-	
 
 	return 0;
 }
