+++ RRTracerUnitTests/RRT_Canvas_UnitTests.cpp
@@ -1,18 +1,18 @@
 #include "pch.h"
 #include "CppUnitTest.h"
 
-//#include "../RRTracer/Canvas.hpp"
+#include "../RRTracer/Canvas.h"
 
 using namespace Microsoft::VisualStudio::CppUnitTestFramework;
 
 namespace RRT_Canvas_UnitTests
 {	
-	TEST_CLASS(RRT_Canvas_Constructor_Tests)
+	TEST_CLASS(RRT_Canvas_UnitTests)
 	{
 	public:
 		TEST_METHOD(Canvas_Constructor_Test)
 		{
-			//RRT::Canvas canvas = RRT::Canvas(10, 20);
+			
 		}
 	};
 }
